tests/wpt/tests/css/css-gaps/parsing/gap-decorations-color-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decorations: column-rule-color parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-gaps-1/#column-row-rule-color">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="*-rule-color supports only the grammar '[ <line-color-list> | <auto-line-color-list> ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
const properties = ["column-rule-color", "row-rule-color", "rule-color"];
for (let property of properties) {
    test_invalid_value(property, "auto");

    test_invalid_value(property, "red, blue, green");
    test_invalid_value(property, "repeat(auto, red, blue, green)");
    test_invalid_value(property, "repeat(0, red, blue, green)");
    test_invalid_value(property, "repeat(-1, red, blue, green)");
    test_invalid_value(property, "repeat(auto, red) red repeat(auto, blue)");
}
</script>
</body>
</html>
